requirements.txt
streamlit>=1.36.0
pypdf>=4.2.0
python-dotenv>=1.0.1
chromadb>=0.5.5
sentence-transformers>=3.0.1
langchain-text-splitters>=0.2.0
openai>=1.51.0
pdfplumber>=0.7.6
faiss-cpu>=1.7.4
pandas>=2.0.0
numpy>=1.26.0
